moved vehicle to separate dir
